+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 
 - Fixed intermittent IR learning errors by properly owning and copying received RMT symbol data before passing from ISR to task.
 - Ignore self when discovering peers for dual comm
+- Prevent crash and spam in EAPOL Logoff attack
 
 ## Revival v1.7.2